Added interruptible sleep primitive
To be able to efficiently debounce using pthreads, i.e., block the
maximum amount of time in the kernel instead of waiting busily, we'll
need a sleep primitive that we can wake from a different thread. This is
the isleep() primitive. It's a very simple wrapper around
cond_wait/cond_broadcast.