Some bugfixes and our first patterns
My wife and I knitted some first patterns and ironed out some bugs.
First, the pattern is now inverted when sending to the machine, but it's
displayed the way it's drawn. Then there was a bug with the solenoids
not being reset after the last line, which probably is fixed now
(untested). Also added a logo (the one we knitted) and some error
handling.