Implemented significant portion of the UI
UI is now coming together. Setting the modes in the knitserver works,
repeat/on-off is implemented and the indicator bar that shows the active
row also works nicely. knitserver still doesn't do any actuation, but
that should be comparatively easy. Also got the events to work by
choosing a gevent-based asyncio loop. This allows the delivery of
Websocket data when through a POST REST call data is modified. Note that
this doesn't work with my stock Ubuntu uwsgi, I had to use the pip3
version.