Some more UI improvements
Added a manual mode to the UI and a clear indication of when the
carriage position was locked. Fixed a bug that any movement of the
carriage would cause it to think it's calibrated. Added a "pattern
finished" sound (although that's played even when manually disabling the
knitting currently, so that needs fixing).